quoting_style: move EscapeChar stuff to its own file
